+++ src/components/GuideListView/_guide-list-view.scss
@@ -1,49 +1,33 @@
 @import '../../styles/responsive';
 
-#guide-list {
+.guide-list {
+  padding: 8px;
   display: flex;
-  justify-content: center;
+  // flex-direction: column;
+  align-items: center;
   overflow: hidden;
-  flex: 1;
-  .guide-list {
-    padding: 8px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    overflow: hidden;
+  // width: 100%;
+  max-width: 800px;
+
+  @include respond-to(gt-md) {
+    margin-left: 258px;
+  }
+
+  &__content {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, 100%);
+    grid-gap: 10px;
     width: 100%;
-    max-width: 800px;
+    justify-content: center;
+    overflow: auto;
+    padding-top: 8px;
 
-    @include respond-to(gt-md) {
-      margin-left: 258px;
+    .euiFlexItem {
+      margin-bottom: 0 !important;
     }
 
-    &__content {
-      display: grid;
-      grid-template-columns: repeat(auto-fill, 100%);
-      grid-gap: 10px;
-      width: 100%;
-      justify-content: center;
-      overflow: auto;
-      padding-top: 8px;
-
-      .euiFlexItem {
-        margin-bottom: 0 !important;
-      }
-
-      // @include respond-to(gt-xs) {
-      //   grid-template-columns: repeat(auto-fill, 400px);
-      // }
-    }
-
-    &__twitch {
-      display: none;
-
-      @include respond-to(gt-md) {
-        display: block;
-        margin-top: 70px;
-        margin-left: 4px;
-      }
-    }
+    // @include respond-to(gt-xs) {
+    //   grid-template-columns: repeat(auto-fill, 400px);
+    // }
   }
 }
