+++ src/components/HomePageView/_home-page-view.scss
@@ -1,0 +1,32 @@
+@import '../../styles/responsive';
+
+#home-view {
+  display: flex;
+  justify-content: center;
+  overflow: hidden;
+  flex: 1;
+  @include respond-to(xs) {
+    .search-controls {
+      padding: 8px;
+    }
+  }
+
+  .home-view {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    &__twitch {
+      display: none;
+
+      @include respond-to(gt-md) {
+        display: block;
+        margin-top: 70px;
+        margin-left: 4px;
+      }
+    }
+
+    @include respond-to(gt-md) {
+      margin-left: 258px;
+    }
+  }
+}
